Figure out what's breaking it

+++ crates/epaint/src/text/text_layout_types.rs
@@ -345,8 +345,8 @@ pub struct TextWrapping {
     /// eliding the text after [`Self::max_width`] is reached.
     /// When you set `max_rows = 1`, it is recommended you also set [`Self::break_anywhere`] to `true`.
     ///
-    /// Default value: `usize::MAX`.
-    pub max_rows: usize,
+    /// Default value: `u32::MAX`.
+    pub max_rows: u32,
 
     /// If `true`: Allow breaking between any characters.
     /// If `false` (default): prefer breaking between words, etc.
